Graceful transition without stopping
Graceful transition without stopping.

Smooth simultaneous transition of both amplitude and frequency.
Amplitude is changed in set time, but frequency may change faster or
slower than commanded.

That's because f. is only changed when motor reveres its direction
during oscillation. Since f. is not constant during transition, the
number of reversal and amount of f. change at each reversal is a
function of starting and stopping frequency which is not calculated
exactly in code. Approximation that gives good results is used.